+++ docs/search/functions_a.js
@@ -2,5 +2,6 @@ var searchData=
 [
   ['loadmigrations',['loadMigrations',['../classBristolSU_1_1Support_1_1Module_1_1ModuleServiceProvider.html#a0b4cc1cd93a8b14f435f426070d38a94',1,'BristolSU::Support::Module::ModuleServiceProvider']]],
   ['logic',['logic',['../classBristolSU_1_1Support_1_1Filters_1_1FilterInstance.html#a9a3d4f0fc68ddf32d6a0f0a282b183c5',1,'BristolSU\Support\Filters\FilterInstance\logic()'],['../classBristolSU_1_1Support_1_1Permissions_1_1Models_1_1ModuleInstancePermission.html#a08fc219c42e6ae8011af2dd7208cffd0',1,'BristolSU\Support\Permissions\Models\ModuleInstancePermission\logic()']]],
-  ['logictester',['logicTester',['../classBristolSU_1_1Support_1_1Testing_1_1FakesLogicTesters.html#abbcc6972bd5671538fdae76f3b89ab29',1,'BristolSU::Support::Testing::FakesLogicTesters']]]
+  ['logictester',['logicTester',['../classBristolSU_1_1Support_1_1Testing_1_1FakesLogicTesters.html#abbcc6972bd5671538fdae76f3b89ab29',1,'BristolSU::Support::Testing::FakesLogicTesters']]],
+  ['logout',['logout',['../interfaceBristolSU_1_1Support_1_1User_1_1Contracts_1_1UserAuthentication.html#a75f8ce299a4d8f3d4d7d1ac6773919a3',1,'BristolSU\Support\User\Contracts\UserAuthentication\logout()'],['../classBristolSU_1_1Support_1_1User_1_1UserApiAuthentication.html#a024516f9cec5aa8106dbdc1ee87e8768',1,'BristolSU\Support\User\UserApiAuthentication\logout()'],['../classBristolSU_1_1Support_1_1User_1_1UserWebAuthentication.html#a127a6e239ae326d1447daf619f590c8f',1,'BristolSU\Support\User\UserWebAuthentication\logout()']]]
 ];