+++ src/User/Contracts/UserRepository.php
@@ -1,10 +1,9 @@
 <?php
 
-
 namespace BristolSU\Support\User\Contracts;
 
-
 use BristolSU\Support\User\User;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Support\Collection;
 
 /**
@@ -14,37 +13,33 @@ interface UserRepository
 {
 
     /**
-     * Get a user where their identity matches the argument. An identity can be qualified as a student ID or email
-     * 
-     * @param string $identity Student ID or email address of the user
-     * @return User|null
-     */
-    public function getWhereIdentity($identity);
-
-    /**
-     * Get a user matching the given email address
+     * Get a user matching the given email address.
      * 
      * @param string $email Email address of the user
-     * @return User|null
+     * @return User
      */
-    public function getWhereEmail($email);
+    public function getWhereEmail($email): User;
+    
+    /**
+     * Get a user matching the given control ID
+     *
+     * @param int $controlId Control ID of the user
+     * @return User
+     */
+    public function getFromControlId(int $controlId): User;
 
     /**
-     * Create a user. 
-     * 
+     * Create a user.
+     *
      * Attributes should be those in the database
      * [
-     *      'forename' => 'Forename',
-     *      'surname' => 'Surname',
-     *      'email' => 'email@example.com',
-     *      'student_id' => 'student id',
-     *      'control_id' => 1 // ID of the control user model representing the user
-     * ]
-     * 
+     *      'control_id' => 1, // ID of the control user model representing the user
+     * ];
+     *
      * @param array $attributes Attributes to create the user with
      * @return User
      */
-    public function create(array $attributes);
+    public function create(array $attributes): User;
 
     /**
      * Get all users registered in the database
@@ -52,5 +47,14 @@ interface UserRepository
      * @return User[]|Collection
      */
     public function all();
+
+    /**
+     * Get a user by remember token
+     * 
+     * @param string $token Remember token
+     * @return User
+     * @throws ModelNotFoundException
+     */
+    public function getFromRememberToken(string $token): User;
 
 }