+++ CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0] - (04/02/2020)
+
+### Added
+- doctrine/dbal dependency for editing migrations
+- Logout function for UserAuthentication
+- UserRepository::getWhereEmail function to retrieve a user via email
+- UserRepository::getFromControlId function to retrieve a user via control ID
+- UserRepository::getFromRememberToken function to get by remember token
+- Functions to get the control user from the database user
+
+### Removed
+- Forename, surname, email and student ID columns from database user model.
+- UserRepository::getWhereIdentity function
+
 ## [1.1] - (30/01/2020)
 
 ### Changed
@@ -35,7 +49,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial Release
 
-[Unreleased]: https://github.com/bristol-su/support/compare/v1.0.1...HEAD
+[Unreleased]: https://github.com/bristol-su/support/compare/v2.0...HEAD
+[2.0]: https://github.com/bristol-su/support/compare/v1.1...v2.0
 [1.1]: https://github.com/bristol-su/support/compare/v1.0.1...v1.1
 [1.0.1]: https://github.com/bristol-su/support/compare/v1.0...v1.0.1
 [1.0]: https://github.com/bristol-su/support/releases/tag/v1.0